--- packages/extension/src/view/popup/stories/data.ts
@@ -1,2631 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-export const dummyData = {
-  '2141247110': {
-    cookies: {
-      '1P_JAR': {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'These cookies are set via embedded youtube-videos. They register anonymous statistical data on for example how many times the video is displayed and what settings are used for playback.',
-          domain: '.gstatic.com',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: '1P_JAR',
-          platform: 'Google',
-          retention: '1 month',
-          wildcard: '0',
-        },
-        frameIdList: [12672, 0, 12634],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.google.com',
-          expires: '2023-09-13T03:11:40.052Z',
-          httponly: false,
-          name: '1P_JAR',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '2023-08-14-03',
-        },
-        url: 'https://www.google.com/recaptcha/api2/aframe',
-      },
-      APC: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0, 12692, 12655],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'doubleclick.net',
-          expires: '2024-02-10T03:15:13.000Z',
-          httponly: false,
-          name: 'APC',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'AfxxVi4q71Kniix5Sds0ANfZ8FI9NnXoEa4OKDoziBMxO163RVmrqQ',
-        },
-        url: 'https://googleads.g.doubleclick.net/pagead/interaction/?ai=&sigh=BpnfxIaauQU&cid=CAQSTgBpAlJWJhG957iiPQM-u9Qi61mWPR6eFTTarfv0pqg0yxpObU1_f5mCWO4La9fQkQf2SlgLdpmUB0N0XlWgQmSnfEk9qN-Hfek6Nz3Jgg&label=window_focus&gqid&qqid=COK0kMOX24ADFZKjZgIdQBMBDQ&fg=1',
-      },
-      AWSALB: {
-        analytics: {
-          category: 'Functional',
-          dataController: 'Amazon Web Services',
-          description:
-            'These cookies enable us to allocate server traffic to make the user experience as smooth as possible. A so-called load balancer is used to determine which server currently has the best availability. The information generated cannot identify you as an individual.',
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://aws.amazon.com/privacy/',
-          name: 'AWSALB',
-          platform: 'Amazon Web Services',
-          retention: 'Session',
-          wildcard: '0',
-        },
-        frameIdList: [0, 12667, 12631],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.dotmetrics.net',
-          expires: '2023-08-21T03:15:13.000Z',
-          httponly: false,
-          name: 'AWSALB',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value:
-            'TlsPHlF6mIBy/meNvNArSCeTmalCYt6cjCGFYhJlUIXgRz7smiFFaUUEf9Z+/qjpq0WhQcu965eJzFAXcXOPIj5t0j4QOPIC/SaViKopv51L9w/5sY5/rMlj4jWZ',
-        },
-        url: 'https://uk-script.dotmetrics.net/BeaconEvent.dotmetrics?1691982913096',
-      },
-      AWSALBCORS: {
-        analytics: {
-          category: 'Functional',
-          dataController: 'Amazon Web Services',
-          description:
-            'For continued stickiness support with CORS use cases after the Chromium update, we are creating additional stickiness cookies for each of these duration-based stickiness features named AWSALBCORS (ALB).',
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://aws.amazon.com/privacy/',
-          name: 'AWSALBCORS',
-          platform: 'Amazon Web Services',
-          retention: 'Session',
-          wildcard: '0',
-        },
-        frameIdList: [0, 12667, 12631],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'uk-script.dotmetrics.net',
-          expires: '2023-08-21T03:15:13.000Z',
-          httponly: false,
-          name: 'AWSALBCORS',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value:
-            'TlsPHlF6mIBy/meNvNArSCeTmalCYt6cjCGFYhJlUIXgRz7smiFFaUUEf9Z+/qjpq0WhQcu965eJzFAXcXOPIj5t0j4QOPIC/SaViKopv51L9w/5sY5/rMlj4jWZ',
-        },
-        url: 'https://uk-script.dotmetrics.net/BeaconEvent.dotmetrics?1691982913096',
-      },
-      AWSALBTG: {
-        analytics: {
-          category: 'Functional',
-          dataController: 'Amazon Web Services',
-          description:
-            'For continued stickiness support with CORS use cases after the Chromium update, we are creating additional stickiness cookies for each of these duration-based stickiness features named AWSELBCORS (ALB).',
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://aws.amazon.com/privacy/',
-          name: 'AWSALBTG',
-          platform: 'Amazon Web Services',
-          retention: 'Session',
-          wildcard: '0',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.the-ozone-project.com',
-          expires: '2023-08-21T03:15:15.000Z',
-          httponly: false,
-          name: 'AWSALBTG',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value:
-            'LukfeGcnlwZKn7Lg3EWrkA1eGS4/AC8mbhBLqIfXu01NpCEMbbXXPywCuzAoi6S2pRfeVRtqWX/Gxp7c2pk+6YeU6PpWrU3y5+H98ugrnLvTLa8R7PAOtEh9OwMs7HeRf+APijK5gpGe0uWxFH+6EKRUOv2dWx1eesVvPeUeCwROtt0p2tE=',
-        },
-        url: 'https://elb.the-ozone-project.com/cookie_sync',
-      },
-      AWSALBTGCORS: {
-        analytics: {
-          category: 'Functional',
-          dataController: 'Amazon Web Services',
-          description:
-            'For continued stickiness support with CORS use cases after the Chromium update, we are creating additional stickiness cookies for each of these duration-based stickiness features named AWSELBCORS (ALB).',
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://aws.amazon.com/privacy/',
-          name: 'AWSALBTGCORS',
-          platform: 'Amazon Web Services',
-          retention: 'Session',
-          wildcard: '0',
-        },
-        frameIdList: [12705, 12662, 0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'elb.the-ozone-project.com',
-          expires: '2023-08-21T03:15:15.000Z',
-          httponly: false,
-          name: 'AWSALBTGCORS',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value:
-            'LukfeGcnlwZKn7Lg3EWrkA1eGS4/AC8mbhBLqIfXu01NpCEMbbXXPywCuzAoi6S2pRfeVRtqWX/Gxp7c2pk+6YeU6PpWrU3y5+H98ugrnLvTLa8R7PAOtEh9OwMs7HeRf+APijK5gpGe0uWxFH+6EKRUOv2dWx1eesVvPeUeCwROtt0p2tE=',
-        },
-        url: 'https://elb.the-ozone-project.com/cookie_sync',
-      },
-      C: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Adform',
-          description:
-            'Used to determine if browser of user accepts cookies or not',
-          domain: 'adform.net (3rd party)',
-          gdprUrl: 'https://site.adform.com/privacy-center/overview',
-          name: 'C',
-          platform: 'Adform',
-          retention: '60 days till 3650 days',
-          wildcard: '0',
-        },
-        frameIdList: [12555],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.adform.net',
-          expires: '2023-08-19T19:30:30.745Z',
-          httponly: false,
-          name: 'C',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1',
-        },
-        url: 'https://a1.adform.net/Serving/Event/?bn=64971431&event=141&time=1086&baid=59137935&asid=12902552&name=Mouse%20Over%20on%20Banner&imprid=2333833189732822627&icid=1012855606482017444&eData=qj8KvLbm2SuXrpCUTL9M3m4864aMyu4tsrd21LaNZhMJDwKV3Zer3KUHcP2LzeMn1qRzCmRYE4frsyoG_hpNj96vWmW1dlSa0&rnd=641077662',
-      },
-      DM_SitId1778: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: 'www.bbc.com',
-          expires: 0,
-          httponly: false,
-          name: 'DM_SitId1778',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value: '1',
-        },
-        url: 'https://www.bbc.com/worklife/article/20230731-the-great-resignation-is-over-what-does-that-mean',
-      },
-      DM_SitId1778SecId13934: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: 'www.bbc.com',
-          expires: 0,
-          httponly: false,
-          name: 'DM_SitId1778SecId13934',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value: '1',
-        },
-        url: 'https://www.bbc.com/worklife/article/20230731-the-great-resignation-is-over-what-does-that-mean',
-      },
-      DSID: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'This cookie is used for targeting, analyzing and optimisation of ad campaigns in DoubleClick/Google Marketing Suite',
-          domain: 'doubleclick.net (3rd party)',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: 'DSID',
-          platform: 'DoubleClick/Google Marketing',
-          retention: '2 weeks',
-          wildcard: '0',
-        },
-        frameIdList: [0, 12692, 12655],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.doubleclick.net',
-          expires: '2023-08-17T00:00:00.360Z',
-          httponly: true,
-          name: 'DSID',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'AGnTgWf_dxKTj8k9w2DGALBYEQnqiF4WWFjbNgJIry89RsVU9JU041onrOda0W42dr2JQPyiEBQkqfwy22D7KM0fDKDq98rspy8uYC_Jo0NCiCRNHuI27nVEKggzgMlWzPlwMyuZmpLTrBfwHD3Iye27wEV-qWham7G5rRvbWeW3wBacCqKlBJeI3tER-LgvYcDRZG8rV2L5yPjRXsNVzd7WXYqJ4VHfz7IpZPLz5xdQ_sHYBHuyu1D_7x3bkK8Fdld8sNJhrw_72I2eg2ZDv_qzXRxQ4FvVmhqD48C6gnKk_8IwGCbpohM',
-        },
-        url: 'https://googleads.g.doubleclick.net/pagead/interaction/?ai=&sigh=BpnfxIaauQU&cid=CAQSTgBpAlJWJhG957iiPQM-u9Qi61mWPR6eFTTarfv0pqg0yxpObU1_f5mCWO4La9fQkQf2SlgLdpmUB0N0XlWgQmSnfEk9qN-Hfek6Nz3Jgg&label=window_focus&gqid&qqid=COK0kMOX24ADFZKjZgIdQBMBDQ&fg=1',
-      },
-      'DotMetrics.DeviceKey': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.dotmetrics.net',
-          expires: '2024-08-13T03:15:13.000Z',
-          httponly: false,
-          name: 'DotMetrics.DeviceKey',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value:
-            'DeviceID=141997162fcea6b68fca8650329493c01a9429c9e58d6fde1a378e59af4e4861',
-        },
-        url: 'https://uk-script.dotmetrics.net/BeaconEvent.dotmetrics?1691982913096',
-      },
-      'DotMetrics.SessionCookieTemp': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.dotmetrics.net',
-          expires: 0,
-          httponly: false,
-          name: 'DotMetrics.SessionCookieTemp',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value: 'SessionId=a9fa9355-4b7f-4605-bc99-518c866de79f',
-        },
-        url: 'https://uk-script.dotmetrics.net/BeaconEvent.dotmetrics?1691982913096',
-      },
-      'DotMetrics.SessionCookieTempTimed': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.dotmetrics.net',
-          expires: '2023-08-14T03:45:13.000Z',
-          httponly: false,
-          name: 'DotMetrics.SessionCookieTempTimed',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value: 'SessionId=a9fa9355-4b7f-4605-bc99-518c866de79f',
-        },
-        url: 'https://uk-script.dotmetrics.net/BeaconEvent.dotmetrics?1691982913096',
-      },
-      'DotMetrics.UniqueUserIdentityCookie': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.dotmetrics.net',
-          expires: '2024-08-13T03:15:13.000Z',
-          httponly: false,
-          name: 'DotMetrics.UniqueUserIdentityCookie',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value:
-            'UserID=72cdbafd-13e9-4c5d-81aa-e475dd3da3b5&Created=08/02/2023 11:02:59&UserMode=2&guid=2aab4c2c-5189-4b58-850a-808961da37d9&ver=1',
-        },
-        url: 'https://uk-script.dotmetrics.net/BeaconEvent.dotmetrics?1691982913096',
-      },
-      IDE: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'This cookie is used for targeting, analyzing and optimisation of ad campaigns in DoubleClick/Google Marketing Suite',
-          domain: 'doubleclick.net (3rd party)',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: 'IDE',
-          platform: 'DoubleClick/Google Marketing',
-          retention: '2 years',
-          wildcard: '0',
-        },
-        frameIdList: [0, 12692, 12655],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.doubleclick.net',
-          expires: '2024-09-17T00:21:59.583Z',
-          httponly: true,
-          name: 'IDE',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'AHWqTUk1HQDSPPwKkeGwqi7vGrJF8jQBBf_aJPX5HaOYBZsyS0PvS5FsyuGTe0rrHEs',
-        },
-        url: 'https://googleads.g.doubleclick.net/pagead/interaction/?ai=&sigh=BpnfxIaauQU&cid=CAQSTgBpAlJWJhG957iiPQM-u9Qi61mWPR6eFTTarfv0pqg0yxpObU1_f5mCWO4La9fQkQf2SlgLdpmUB0N0XlWgQmSnfEk9qN-Hfek6Nz3Jgg&label=window_focus&gqid&qqid=COK0kMOX24ADFZKjZgIdQBMBDQ&fg=1',
-      },
-      NID: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'This cookies is used to collect website statistics and track conversion rates and Google ad personalisation',
-          domain: 'google.com',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: 'NID',
-          platform: 'Google',
-          retention: '1 year',
-          wildcard: '0',
-        },
-        frameIdList: [12672, 0, 12634],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.google.com',
-          expires: '2024-02-13T02:42:52.882Z',
-          httponly: true,
-          name: 'NID',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            '511=K9solb0VvA-yks9F9Z6UC9vCckKwB515w8xGjPr7WkQf_hQX_juwLVl1tJdcoRAfzNATXtBeLZ-UJCf42xP5NNr_SogvdJzXgIWJHYwHQTgGzvPVg0quyqtTgwMgYrPuf5yxq5ceW6U99FzJ47TYdvgfeTGDzvL2lCcG5WbOJ1bdvyRYuHD43txrhAs6RwvtAwbbKd0ej0Huum5toX1rcwSK-3oF_GhL-dBNXH2qNtWnpDimZDwcpOqB8dGruIQeRo41BDLCV7pH88XgBDz08ex6mOyQWSoF354_4vo25xKpHqMrpTAawgl4uAcdN0Y0j8F2PwYAHXInfe-imVhw6neWoJXti8Kh97a0Oq_LUXeh0-EJEJunGj56TA',
-        },
-        url: 'https://www.google.com/recaptcha/api2/aframe',
-      },
-      TPC: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Adform',
-          description:
-            'Used to determine if browser of user accepts third party cookies or not',
-          domain: 'adform.net (3rd party)',
-          gdprUrl: 'https://site.adform.com/privacy-center/overview',
-          name: 'TPC',
-          platform: 'Adform',
-          retention: '14 days',
-          wildcard: '0',
-        },
-        frameIdList: [12555],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.adform.net',
-          expires: '2023-08-21T02:56:19.297Z',
-          httponly: false,
-          name: 'TPC',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1691981779385',
-        },
-        url: 'https://a1.adform.net/Serving/Event/?bn=64971431&event=141&time=1086&baid=59137935&asid=12902552&name=Mouse%20Over%20on%20Banner&imprid=2333833189732822627&icid=1012855606482017444&eData=qj8KvLbm2SuXrpCUTL9M3m4864aMyu4tsrd21LaNZhMJDwKV3Zer3KUHcP2LzeMn1qRzCmRYE4frsyoG_hpNj96vWmW1dlSa0&rnd=641077662',
-      },
-      UID: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.scorecardresearch.com',
-          expires: '2025-08-03T03:15:09.359Z',
-          httponly: false,
-          name: 'UID',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value: '1820ef102430a460145342d1690974245',
-        },
-        url: 'https://sb.scorecardresearch.com/b?c1=2&c2=19293874&cs_it=b9&cv=4.0.0%2B2301240627&ns__t=1691982909229&ns_c=UTF-8&cs_ucfr=1&gdpr=0&gdpr_p1t=&gdpr_li=&gdpr_purps=&gdpr_pcc=&cs_cmp_nc=0&cs_cmp_id=6&cs_cmp_sv=1&cs_cmp_rt=0&c7=https%3A%2F%2Fwww.bbc.com%2Fworklife%2Farticle%2F20230731-the-great-resignation-is-over-what-does-that-mean&c8=The%20Great%20Resignation%20is%20%27over%27.%20What%20does%20that%20mean%3F%20-%20BBC%20Worklife&c9=',
-      },
-      '__Secure-3PAPISID': {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'Profiles the interests of website visitors to serve relevant and personalised ads through retargeting.',
-          domain: 'google.com',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: '__Secure-3PAPISID',
-          platform: 'Google Ads',
-          retention: '2 years',
-          wildcard: '0',
-        },
-        frameIdList: [12672, 0, 12634],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.google.com',
-          expires: '2024-09-08T11:35:00.123Z',
-          httponly: false,
-          name: '__Secure-3PAPISID',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'gDT4BJmE1x5TgZaJ/AYjfecOyTTaBwOe_2',
-        },
-        url: 'https://www.google.com/recaptcha/api2/aframe',
-      },
-      '__Secure-3PSID': {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'Targeting cookie. Used to profile the interests of website visitors and display relevant and personalised Google ads.',
-          domain: 'google.com',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: '__Secure-3PSID',
-          platform: 'Google Ads',
-          retention: '2 years',
-          wildcard: '0',
-        },
-        frameIdList: [12672, 0, 12634],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.google.com',
-          expires: '2024-09-08T11:35:00.123Z',
-          httponly: true,
-          name: '__Secure-3PSID',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'ZggdI8E3Y6CuAGj7R4s3ijYDMaUXKHhmHxiKSku6Z2JswI1LAhiXsjNsgc6s-0mJFIjhng.',
-        },
-        url: 'https://www.google.com/recaptcha/api2/aframe',
-      },
-      '__Secure-3PSIDCC': {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'Targeting cookie. Used to create a user profile and display relevant and personalised Google Ads to the user.',
-          domain: 'google.com',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: '__Secure-3PSIDCC',
-          platform: 'Google Ads',
-          retention: '2 years',
-          wildcard: '0',
-        },
-        frameIdList: [12672, 0, 12634],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.google.com',
-          expires: '2024-08-13T03:14:39.728Z',
-          httponly: true,
-          name: '__Secure-3PSIDCC',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'APoG2W8ZEhaFXDeHNwZXo9z3uglHEUj-jOZHXI4XiPk8IntI1-k5IOB4e1XcyONbv-Gv5DowUbRo',
-        },
-        url: 'https://www.google.com/recaptcha/api2/aframe',
-      },
-      __cf_bm: {
-        analytics: {
-          category: 'Functional',
-          dataController: 'Cloudflare',
-          description:
-            "Cloudflare's bot products identify and mitigate automated traffic to protect your site from bad bots. Cloudflare places the __cf_bm cookie on End User devices that access Customer sites that are protected by Bot Management or Bot Fight Mode. The __cf_bm cookie is necessary for the proper functioning of these bot solutions.",
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://www.cloudflare.com/privacypolicy/',
-          name: '__cf_bm',
-          platform: 'Cloudflare',
-          retention: 'session',
-          wildcard: '0',
-        },
-        frameIdList: [12705, 0, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.the-ozone-project.com',
-          expires: '2023-08-14T03:44:42.378Z',
-          httponly: true,
-          name: '__cf_bm',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'qErZD2abaYkuxhHJjd8HlZLgGqlP2m.Pj6aBj5MSgRk-1691982882-0-AQBZwLh68E5bU2t9slDrJ1UBrD8Roljrx76rzSQM0tTAvSwF6EewHLTH5r6bwO17PcxS5wcKmXbYKOtPojmkhhc=',
-        },
-        url: 'https://elb.the-ozone-project.com/cdn-cgi/rum?',
-      },
-      __gads: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'This cookie is used by Google for a variety of purposes (e.g., ensuring Frequency Caps work correctly). It includes AdSense if you have AdSense enabled. This cookie is associated with the DoubleClick for Publishers service from Google. Its purpose is to monitor the showing of advertisements on the site, for which the owner may earn some revenue. The main purpose of this cookie is targeting/advertising.',
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: '__gads',
-          platform: 'DoubleClick/Google Marketing',
-          retention: 'various',
-          wildcard: '0',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-08-24T05:13:33.000Z',
-          httponly: false,
-          name: '__gads',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value:
-            'ID=aa5d7652436b0117:T=1690780413:RT=1691982878:S=ALNI_MYOpD40IPQxvBOvaFBOrdKXh58Mxw',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      __gpi: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Google',
-          description:
-            'Collects information on user behaviour on multiple websites. This information is used in order to optimize the relevance of advertisement on the website.',
-          domain: "Advertiser's website domain (1st party)",
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: '__gpi',
-          platform: 'Google Adsense',
-          retention: '13 months',
-          wildcard: '0',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-07T03:15:11.000Z',
-          httponly: false,
-          name: '__gpi',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value:
-            'UID=00000c2c445cdb98:T=1691982911:RT=1691982911:S=ALNI_Mb_saTOironpXKmGylAlt4Z3L3pZA',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      __tbc: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-17T01:54:50.395Z',
-          httponly: false,
-          name: '__tbc',
-          path: '/',
-          samesite: 'lax',
-          secure: false,
-          value:
-            '%7Bkpex%7D7_n15eNMtY3kuYrZnrmVp2PKIXkN2LFkErXRAYjOGOADZTCrEaa8DnixAlsSJgqs',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      _cb: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-12T03:15:08.000Z',
-          httponly: false,
-          name: '_cb',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value: 'BGaOpDhtp6YCo79nI',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      _cb_svref: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2023-08-14T03:45:08.000Z',
-          httponly: false,
-          name: '_cb_svref',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value: 'null',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      _chartbeat2: {
-        analytics: {
-          category: 'Analytics',
-          dataController: 'Chartbeat',
-          description:
-            'Cookie is used to register if a person has visited the domain before (to calculate new vs returning users).',
-          domain: 'website domain (1st party)',
-          gdprUrl: 'https://chartbeat.com/privacy/',
-          name: '_chartbeat2',
-          platform: 'Chartbeat',
-          retention: '30 days',
-          wildcard: '0',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-12T03:15:08.000Z',
-          httponly: false,
-          name: '_chartbeat2',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value:
-            '.1690445942438.1691982908315.1111111111110001.hsZi62ArwtCRWL3MDhWSd2B6o1Rf.2',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      _chartbeat4: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2023-08-14T03:16:07.000Z',
-          httponly: false,
-          name: '_chartbeat4',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value: 't=B7_pz6DMzolbBrvXzjDb8NB_p-RF&E=0&x=0&c=0.54&y=5579&w=185',
-        },
-        url: 'https://www.bbc.com/userinfo',
-      },
-      _pbjs_userid_consent_data: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: 'www.bbc.com',
-          expires: '2023-09-13T03:14:38.000Z',
-          httponly: false,
-          name: '_pbjs_userid_consent_data',
-          path: '/',
-          samesite: 'lax',
-          secure: false,
-          value: '3524755945110770',
-        },
-        url: 'https://www.bbc.com/worklife/article/20230731-the-great-resignation-is-over-what-does-that-mean',
-      },
-      _pcid: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-12T01:54:50.000Z',
-          httponly: false,
-          name: '_pcid',
-          path: '/',
-          samesite: 'lax',
-          secure: true,
-          value: '%7B%22browserId%22%3A%22lkqpmxkvbx5jgfto%22%7D',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      _pctx: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-12T01:54:50.000Z',
-          httponly: false,
-          name: '_pctx',
-          path: '/',
-          samesite: 'lax',
-          secure: true,
-          value:
-            '%7Bu%7DN4IgrgzgpgThIC4B2YA2qA05owMoBcBDfSREQpAeyRCwgEt8oBJAEzIGYA2ABi644AmQQFYALDwDsYgIwiR3AJwgAvkA',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      _sp_su: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-07-30T05:13:32.152Z',
-          httponly: false,
-          name: '_sp_su',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'false',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ar_debug: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0, 12692, 12655],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.doubleclick.net',
-          expires: '2023-09-08T13:28:14.774Z',
-          httponly: true,
-          name: 'ar_debug',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1',
-        },
-        url: 'https://googleads.g.doubleclick.net/pagead/interaction/?ai=&sigh=BpnfxIaauQU&cid=CAQSTgBpAlJWJhG957iiPQM-u9Qi61mWPR6eFTTarfv0pqg0yxpObU1_f5mCWO4La9fQkQf2SlgLdpmUB0N0XlWgQmSnfEk9qN-Hfek6Nz3Jgg&label=window_focus&gqid&qqid=COK0kMOX24ADFZKjZgIdQBMBDQ&fg=1',
-      },
-      atuserid: {
-        analytics: {
-          category: 'Analytics',
-          dataController: 'AT Internet',
-          description: 'Visitor ID for client-side cookie sites',
-          domain: "Advertiser's website domain",
-          gdprUrl: 'N/A',
-          name: 'atuserid',
-          platform: 'AT Internet',
-          retention: '13 months by default, modifiable',
-          wildcard: '0',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-14T03:15:08.000Z',
-          httponly: false,
-          name: 'atuserid',
-          path: '/',
-          samesite: 'strict',
-          secure: true,
-          value:
-            '%7B%22name%22%3A%22atuserid%22%2C%22val%22%3A%228639fc19-1990-4cc9-8cac-49915009f3ef%22%2C%22options%22%3A%7B%22end%22%3A%222024-09-14T03%3A15%3A08.340Z%22%2C%22path%22%3A%22%2F%22%7D%7D',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      blaize_session: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2023-12-18T05:13:32.630Z',
-          httponly: false,
-          name: 'blaize_session',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value: '4c213585-5666-4221-aa89-958663f1e728',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      blaize_tracking_id: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-03T05:13:32.630Z',
-          httponly: false,
-          name: 'blaize_tracking_id',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value: '935115ec-cef4-4076-bca3-2774113de44f',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ckns_eds: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.www.bbc.com',
-          expires: '2024-09-04T13:04:42.267Z',
-          httponly: false,
-          name: 'ckns_eds',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value: 'INS-vt29-666188954:923108334-1690895081',
-        },
-        url: 'https://www.bbc.com/worklife/assets/worklife/android-chrome-192x192.png',
-      },
-      ckns_explicit: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-08-14T03:15:08.000Z',
-          httponly: false,
-          name: 'ckns_explicit',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '0',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ckns_mvt: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-07-31T10:11:13.000Z',
-          httponly: false,
-          name: 'ckns_mvt',
-          path: '/',
-          samesite: '',
-          secure: true,
-          value: 'adbe238d-38fb-4958-b25f-6892bc0dad0b',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ckns_policy: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-08-14T03:15:08.000Z',
-          httponly: false,
-          name: 'ckns_policy',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '111',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ckns_policy_exp: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-14T03:15:08.000Z',
-          httponly: false,
-          name: 'ckns_policy_exp',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1723605308253',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ckpf_ppid: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-08-13T03:15:09.000Z',
-          httponly: false,
-          name: 'ckpf_ppid',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value: '6d844e2beb7e447bb25fe518658472ad',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      'data-a': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-10-17T09:05:43.315Z',
-          httponly: false,
-          name: 'data-a',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '3919552754408456762~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-acu': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T04:36:05.822Z',
-          httponly: false,
-          name: 'data-acu',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '99705d3c-6136-49de-970b-d410e92016d0~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-amb': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T09:07:46.353Z',
-          httponly: false,
-          name: 'data-amb',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '3625603391085596777~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-ap': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-08-09T11:09:10.369Z',
-          httponly: false,
-          name: 'data-ap',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'setstatuscode~~3',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-bai': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-21T02:00:19.827Z',
-          httponly: false,
-          name: 'data-bai',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '3b513474a036df683c02fb826d1a9f96~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-bs': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-08-03T08:08:46.955Z',
-          httponly: false,
-          name: 'data-bs',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'a599a964-e0e3-48f7-9af5-1c124a822da9~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-bt': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T09:07:12.885Z',
-          httponly: false,
-          name: 'data-bt',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '289aff8b-6717-4c45-bee0-5a0ae7b32989~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-c': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-09-08T08:17:46.120Z',
-          httponly: false,
-          name: 'data-c',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'k-N3Vnd-77lJdgFgjNJ8Fjv8nY7_m6kd827ghvWA~~3',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-c-ts': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-09-08T08:17:46.120Z',
-          httponly: false,
-          name: 'data-c-ts',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1691569066',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-co': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T03:51:42.290Z',
-          httponly: false,
-          name: 'data-co',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'AAAJrJ9ABBQSxgMojupYAAAAAAA~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-dat': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T03:51:41.430Z',
-          httponly: false,
-          name: 'data-dat',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'setstatuscode~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-doc': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T09:07:15.034Z',
-          httponly: false,
-          name: 'data-doc',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'e9fef13b-338c-4e4f-8cfe-e4e17adeebb5~~3',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-g': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-08-17T12:32:50.159Z',
-          httponly: false,
-          name: 'data-g',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '11~~18',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-le': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T09:07:47.325Z',
-          httponly: false,
-          name: 'data-le',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '4a49b394-26ca-11ee-ad9d-801844df0ab8~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-lop': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-27T02:50:10.061Z',
-          httponly: false,
-          name: 'data-lop',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '08ef69f0-7cc9-46db-829a-fa5dfee1bf6a~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-mf': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T03:51:42.097Z',
-          httponly: false,
-          name: 'data-mf',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'd9a7875b-5e24-42e7-8fb1-099854072aa9~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-mic': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-27T02:50:09.927Z',
-          httponly: false,
-          name: 'data-mic',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'UaontCzPO84TseKhBzCY2Mwqk7R_v87ninR6iljlRzARoskzwH_5YyeG38VBHpKEz3zeedgAtvA~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-o': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T09:05:42.714Z',
-          httponly: false,
-          name: 'data-o',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'a0a74f26-7922-445b-a7fc-e89b3c4d3493~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-ot': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T09:06:44.474Z',
-          httponly: false,
-          name: 'data-ot',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'QlXPWsCPr0SjY97cbA-WXl5IugbPmX8hRFJo8-51WGM~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-p': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T01:51:26.225Z',
-          httponly: false,
-          name: 'data-p',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'uAOynRwGOxlT~~3',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-pba': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T09:06:13.306Z',
-          httponly: false,
-          name: 'data-pba',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '84103F35-4466-4816-B699-CEF4B27C725A~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-pbs': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-08-13T03:15:16.000Z',
-          httponly: false,
-          name: 'data-pbs',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'setstatuscode~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-pri': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-08-04T09:41:50.907Z',
-          httponly: false,
-          name: 'data-pri',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '64c68f9f76864~~34',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-r': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-01-16T09:05:42.950Z',
-          httponly: false,
-          name: 'data-r',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'LKA4BZ27-9-1KRG~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-r1': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T09:05:43.321Z',
-          httponly: false,
-          name: 'data-r1',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'RX-e6117e92-9ce1-4bf5-b8cb-562e5c80b805-004~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-rbh': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T03:52:12.486Z',
-          httponly: false,
-          name: 'data-rbh',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'ycFQc5NMJkJ0z0zJv0NB~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-rk': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T03:51:42.287Z',
-          httponly: false,
-          name: 'data-rk',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1975180292632303609~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-sma': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T04:36:05.845Z',
-          httponly: false,
-          name: 'data-sma',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'a35f2995415f9d74a346842ba61602135a20dc90315cb06c642c7307a0129048~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-so': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-08-17T09:06:41.939Z',
-          httponly: false,
-          name: 'data-so',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '9ee4dacf-cf84-4c41-a9b7-2c9ce12e05df~~35',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-tam': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-08-17T05:49:23.339Z',
-          httponly: false,
-          name: 'data-tam',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'setstatuscode~~35',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-ttd': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2023-08-17T05:48:53.439Z',
-          httponly: false,
-          name: 'data-ttd',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '16740dc8-47c3-4334-8a87-b19478d31baf~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-v': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-21T02:00:18.845Z',
-          httponly: false,
-          name: 'data-v',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '{{APID}}~~1',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-xc': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T09:07:16.541Z',
-          httponly: false,
-          name: 'data-xc',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '212163822799578~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-xu': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-18T03:51:41.666Z',
-          httponly: false,
-          name: 'data-xu',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'MT5IYbk41QmcSB5~~8',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'data-yl': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-08-02T05:50:00.202Z',
-          httponly: false,
-          name: 'data-yl',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'setstatuscode~~3',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      'ecos.dt': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.www.bbc.com',
-          expires: 0,
-          httponly: false,
-          name: 'ecos.dt',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value: '1691982909231',
-        },
-        url: 'https://www.bbc.com/worklife/assets/worklife/android-chrome-192x192.png',
-      },
-      exactag_new_gk: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'm.exactag.com',
-          expires: '2023-10-26T02:50:04.851Z',
-          httponly: true,
-          name: 'exactag_new_gk',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '96cb1e0f5c184b68ba731749e54a199f%7C26.09.2023%2002%3A50%3A04',
-        },
-        url: 'https://m.exactag.com/ai.aspx?extProvId=63&extPu=lh-mindshare&extProvApi=lh_in&extLi=30357894&extCr=191229156&extPm=372875365&gdpr_consent=&gdpr=0',
-      },
-      exactag_new_uk: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'm.exactag.com',
-          expires: '2024-02-10T03:15:14.000Z',
-          httponly: true,
-          name: 'exactag_new_uk',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '54a5894664eb4c62a75c75e50ace4fb6%7C',
-        },
-        url: 'https://m.exactag.com/ai.aspx?extProvId=63&extPu=lh-mindshare&extProvApi=lh_in&extLi=30357894&extCr=191229156&extPm=372875365&gdpr_consent=&gdpr=0',
-      },
-      exactag_new_user: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'm.exactag.com',
-          expires: '2024-01-30T20:12:58.904Z',
-          httponly: true,
-          name: 'exactag_new_user',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            '1830%7C1%7Ce43ec29234fc4cdeabe67cc8%7C01.01.0001%2000%3A00%3A00%7C03.08.2023%2020%3A12%3A58%7Ce43ec29234fc4cdeabe67cc8%7C873470%7C4884%7CFalse',
-        },
-        url: 'https://m.exactag.com/ai.aspx?extProvId=63&extPu=lh-mindshare&extProvApi=lh_in&extLi=30357894&extCr=191229156&extPm=372875365&gdpr_consent=&gdpr=0',
-      },
-      optimizelyEndUserId: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Optimizely',
-          description:
-            "Stores a visitor's unique Optimizely identifier. It's a combination of a timestamp and random number. No other information about you or your visitors is stored inside.",
-          domain: "Advertiser's website domain",
-          gdprUrl: 'https://www.optimizely.com/privacy/',
-          name: 'optimizelyEndUserId',
-          platform: 'Optimizely',
-          retention: '6 months',
-          wildcard: '0',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-08-13T01:40:21.000Z',
-          httponly: false,
-          name: 'optimizelyEndUserId',
-          path: '/',
-          samesite: '',
-          secure: false,
-          value: 'oeu1690780411414r0.9184144086392749',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      ozone_uid: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'The Ozone Project',
-          description:
-            'This cookie contains unique randomly-generated values that enable the Ozone Project to distinguish browsers and mobile devices.',
-          domain: 'the-ozone-project.com',
-          gdprUrl: 'https://www.ozoneproject.com/privacy-matters',
-          name: 'ozone_uid',
-          platform: 'The Ozone Project',
-          retention: '90 days',
-          wildcard: '0',
-        },
-        frameIdList: [12705, 12662, 0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'elb.the-ozone-project.com',
-          expires: '2023-11-12T03:15:15.000Z',
-          httponly: false,
-          name: 'ozone_uid',
-          path: '/',
-          samesite: 'None',
-          secure: true,
-          value: '2T9CKzK3FG0DGDVqGr8TJxw2IRg',
-        },
-        url: 'https://elb.the-ozone-project.com/cookie_sync',
-      },
-      'permutive-id': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-02-14T03:15:09.000Z',
-          httponly: false,
-          name: 'permutive-id',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'c34a60ba-5922-4b19-a9dd-8a0f4147b2ec',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-      session_session: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: 'm.exactag.com',
-          expires: 0,
-          httponly: true,
-          name: 'session_session',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '50f9bc139fb2469eb1eb5a1d',
-        },
-        url: 'https://m.exactag.com/ai.aspx?extProvId=63&extPu=lh-mindshare&extProvApi=lh_in&extLi=30357894&extCr=191229156&extPm=372875365&gdpr_consent=&gdpr=0',
-      },
-      test_cookie: {
-        analytics: {
-          category: 'Functional',
-          dataController: 'Google',
-          description:
-            "This cookie is set by DoubleClick (which is owned by Google) to determine if the website visitor's browser supports cookies.",
-          domain: 'doubleclick.net',
-          gdprUrl: 'https://privacy.google.com/take-control.html',
-          name: 'test_cookie',
-          platform: 'DoubleClick/Google Marketing',
-          retention: '1 year',
-          wildcard: '0',
-        },
-        frameIdList: [0],
-        headerType: 'response',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.doubleclick.net',
-          expires: '2023-08-14T03:30:14.000Z',
-          httponly: true,
-          name: 'test_cookie',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: 'CheckForPermission',
-        },
-        url: 'https://securepubads.g.doubleclick.net/pcs/view?xai=AKAOjstZX8T_zJIWOoZJpcsBgZF2413l05WKB6bOlhKlhw_sE6Awbkp3Jh7M13W8Erj4hgf-AL_pe-0xhEbUQbI1oGP1xii1pnRcnbFdUB_fRch3vfxBHR_p17tVbenuxfDnoXcJhm-XLy5Tm4OsPPLuz8FVK9DCDnf2JcD2rkEWmKwp5jLIpLoSEOG-wskYoFE0VBkdoq_Iybd9_Bz420i-1T8bmlAkavYMGieNERnDz6zJcKRynO3EkojYpk8-lV5xEaXEFkznbdbnRj2tP6JzgMlQEWkVdcffkO5RKusHq9-SEeWTye1Cj-jMIJ0iBg6n8v8VB62V3R_Z3pLNrBzZ_d015hNRIc7MNsJHH5BAE7TPDDVFDcRG_lDHqQh1OJQ&sai=AMfl-YSfExFr1bXjRAjWiUia96LRprOdsMEfx6eIIG3TeVELcGFM2WvSObLOlngmwtRXEppSpklcChsPHqtJ5Th7U-IJncRfnfuXS9b3C4ZVS3-sDV7aQimFrvTsBZ5IrWYAzK8_1bDjwK23EhmEBDOEXDU&sig=Cg0ArKJSzF3KdMQ9xSvoEAE&uach_m=[UACH]&urlfix=1&uach=WyJtYWNPUyIsIjEzLjEuMCIsImFybSIsIiIsIjExNS4wLjU3OTAuMTcwIixbXSwwLG51bGwsIjY0IixbWyJOb3QvQSlCcmFuZCIsIjk5LjAuMC4wIl0sWyJHb29nbGUgQ2hyb21lIiwiMTE1LjAuNTc5MC4xNzAiXSxbIkNocm9taXVtIiwiMTE1LjAuNTc5MC4xNzAiXV0sMF0%3D&adurl=',
-      },
-      tt_exelate: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12692, 12685, 0, 12655, 12646],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.teads.tv',
-          expires: '2023-08-15T02:56:18.565Z',
-          httponly: false,
-          name: 'tt_exelate',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '',
-        },
-        url: 'https://sync.teads.tv/um/report?ssb_provider_id=3&google_nid=teadstv_ab&fp=1',
-      },
-      tt_liveramp: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12692, 12685, 0, 12655, 12646],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.teads.tv',
-          expires: '2023-08-15T02:56:18.565Z',
-          httponly: false,
-          name: 'tt_liveramp',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '',
-        },
-        url: 'https://sync.teads.tv/um/report?ssb_provider_id=3&google_nid=teadstv_ab&fp=1',
-      },
-      tt_neustar: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12692, 12685, 0, 12655, 12646],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.teads.tv',
-          expires: '2023-08-15T02:56:18.565Z',
-          httponly: false,
-          name: 'tt_neustar',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '',
-        },
-        url: 'https://sync.teads.tv/um/report?ssb_provider_id=3&google_nid=teadstv_ab&fp=1',
-      },
-      tt_viewer: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Teads.com',
-          description:
-            'Teads uses a “tt_viewer” cookie to help personalize the video ads you see on our partner websites.',
-          domain: 'teads.com',
-          gdprUrl: 'https://www.teads.com/privacy-policy/',
-          name: 'tt_viewer',
-          platform: 'Teads',
-          retention: '1 year',
-          wildcard: '0',
-        },
-        frameIdList: [12692, 12685, 0, 12655, 12646],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.teads.tv',
-          expires: '2024-07-17T19:39:31.870Z',
-          httponly: false,
-          name: 'tt_viewer',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '18a5aa70-f792-49a3-83f2-e833434fbdd5',
-        },
-        url: 'https://sync.teads.tv/um/report?ssb_provider_id=3&google_nid=teadstv_ab&fp=1',
-      },
-      uid: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Adform',
-          description: 'Contains a unique ID to identify a user',
-          domain: 'adform.net (3rd party)',
-          gdprUrl: 'https://site.adform.com/privacy-center/overview',
-          name: 'uid',
-          platform: 'Adform',
-          retention: '60 days',
-          wildcard: '0',
-        },
-        frameIdList: [12555],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.adform.net',
-          expires: '2023-10-13T02:56:20.102Z',
-          httponly: false,
-          name: 'uid',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '1012855606482017444',
-        },
-        url: 'https://a1.adform.net/Serving/Event/?bn=64971431&event=141&time=1086&baid=59137935&asid=12902552&name=Mouse%20Over%20on%20Banner&imprid=2333833189732822627&icid=1012855606482017444&eData=qj8KvLbm2SuXrpCUTL9M3m4864aMyu4tsrd21LaNZhMJDwKV3Zer3KUHcP2LzeMn1qRzCmRYE4frsyoG_hpNj96vWmW1dlSa0&rnd=641077662',
-      },
-      uids: {
-        analytics: {
-          category: 'Marketing',
-          dataController: 'Admatic',
-          description:
-            'Registers user behaviour and navigation on the website, and any interaction with active campaigns. This is used for optimizing advertisement and for efficient retargeting.',
-          domain: 'admatic.com.tr (3rd party)',
-          gdprUrl: 'http://www.admatic.com.tr/en/privacy-policy.html',
-          name: 'uids',
-          platform: 'Admatic',
-          retention: '3 months',
-          wildcard: '0',
-        },
-        frameIdList: [12705, 0, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.the-ozone-project.com',
-          expires: '2023-11-01T12:26:56.711Z',
-          httponly: false,
-          name: 'uids',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value:
-            'eyJ0ZW1wVUlEcyI6eyJhZGZvcm0iOnsidWlkIjoiMTAxMjg1NTYwNjQ4MjAxNzQ0NCIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjI1OjQ3Ljc0NTY2MTc3N1oiLCJzb3VyY2UiOiJjb29raWUifSwiYWRueHMiOnsidWlkIjoiMzkxOTU1Mjc1NDQwODQ1Njc2MiIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjI1OjQ1LjkwNzg5NTU5MVoiLCJzb3VyY2UiOiJjb29raWUifSwiYXZvY2V0Ijp7InVpZCI6IjBjNmQ0Y2QxLTg4ZDUtNDgwMi04NTc2LTk4MjNiZmJmMTJkYSIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjU4OjE5LjczNTc4NjQzM1oiLCJzb3VyY2UiOiJjb29raWUifSwiYmVlc3dheCI6eyJ1aWQiOiJBQUl2TTA3SmNIZ0FBQ1l1aXBoN2lnIiwiZXhwaXJlcyI6IjIwMjMtMDgtMTZUMTQ6NTg6MjMuMTIyMzIxODE1WiIsInNvdXJjZSI6ImNvb2tpZSJ9LCJncmlkIjp7InVpZCI6ImE1OTlhOTY0LWUwZTMtNDhmNy05YWY1LTFjMTI0YTgyMmRhOSIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjU4OjE4Ljk3MDkxNzM1WiIsInNvdXJjZSI6ImNvb2tpZSJ9LCJpbXByb3ZlZGlnaXRhbCI6eyJ1aWQiOiJjYTZhMjBjNC0wMDc3LTQ2MTUtYjUyNy00OTM2ZGQ4YWUyZWEiLCJleHBpcmVzIjoiMjAyMy0wOC0xNlQxNDo1ODoyMy45NDQ4MjU3MTFaIiwic291cmNlIjoiY29va2llIn0sIml4Ijp7InVpZCI6IlpMZzUxME1yWHNpNnpTUGsyVnRBS3dBQVx1MDAyNjU0MTYiLCJleHBpcmVzIjoiMjAyMy0wOC0xNlQxNDoyNTo0NS4wMzQ4Mjc1OTZaIiwic291cmNlIjoiY29va2llIn0sImthcmdvIjp7InVpZCI6ImNmNjkyYzFhLWYzMTgtNzI3Ni1jOGU2LTlhMTA1NDhiNGYzZiIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjU4OjIyLjExNTE5ODQ1NFoiLCJzb3VyY2UiOiJjb29raWUifSwib3BlbngiOnsidWlkIjoiNTE1NWE5NzYtNmJiZC00ZDUyLWFmM2YtNTZlMDBmZjk2NGMyIiwiZXhwaXJlcyI6IjIwMjMtMDgtMTZUMTQ6NTg6MjIuMzY0ODY2OTI3WiIsInNvdXJjZSI6ImNvb2tpZSJ9LCJwdWJtYXRpYyI6eyJ1aWQiOiI4NDEwM0YzNS00NDY2LTQ4MTYtQjY5OS1DRUY0QjI3QzcyNUEiLCJleHBpcmVzIjoiMjAyMy0wOC0xN1QxMjoyNjo1Ni43OTIxOTU0MTdaIiwic291cmNlIjoiY29va2llIn0sInJ1Ymljb24iOnsidWlkIjoiTEtBNEJaMjctOS0xS1JHIiwiZXhwaXJlcyI6IjIwMjMtMDgtMTZUMTQ6NTg6MjAuOTQ2NTYyMTU3WiIsInNvdXJjZSI6ImNvb2tpZSJ9LCJzaGFyZXRocm91Z2giOnsidWlkIjoiZTIyYTBhZGQtNjIwOS00NTYzLThhMWUtNDhkMjEwYjY0NDBmIiwiZXhwaXJlcyI6IjIwMjMtMDgtMTZUMTQ6NTg6MjMuNDg1NTQ1OTU1WiIsInNvdXJjZSI6ImNvb2tpZSJ9LCJzbWFydCI6eyJ1aWQiOiIzNTEyMDQzOTgzMDM3MDY3OTc1IiwiZXhwaXJlcyI6IjIwMjMtMDgtMTZUMTQ6MjU6NDcuMTIzNzQ1OTIzWiIsInNvdXJjZSI6ImNvb2tpZSJ9LCJzb3ZybiI6eyJ1aWQiOiJIQWh5VUJaSEMwNFN6a045UmJDc3JiNkkiLCJleHBpcmVzIjoiMjAyMy0wOC0xNlQxNDo1ODoyMC43MjI4ODQwNTFaIiwic291cmNlIjoiY29va2llIn0sInRyaXBsZWxpZnQiOnsidWlkIjoiMjYyMTk4MTk4MzEyMTU4OTI4ODQxMCIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjU4OjE3LjY0NzQ5NTU3MVoiLCJzb3VyY2UiOiJjb29raWUifSwidHRkIjp7InVpZCI6IjE2NzQwZGM4LTQ3YzMtNDMzNC04YTg3LWIxOTQ3OGQzMWJhZiIsImV4cGlyZXMiOiIyMDIzLTA4LTE2VDE0OjI1OjQ2LjcxODA5OTYyOFoiLCJzb3VyY2UiOiJjb29raWUifSwieWFob29zc3AiOnsidWlkIjoieS1QbkJ6M210RTJ1SENHVnhKcWNSdVZHeWNKanc3dW04eGw4S1EwSzRrNGphcTU2cUVRX3RrdHhxdWp5QzlwakVkTkkxYnVqMFVWRVktfkEiLCJleHBpcmVzIjoiMjAyMy0wOC0xNlQxNDo1ODoxOC4wMTYwMjIxODFaIiwic291cmNlIjoiY29va2llIn0sInlpZWxkbW8iOnsidWlkIjoiZzM4NjQxMDYxZTg5Zjk5N2I0OWMiLCJleHBpcmVzIjoiMjAyMy0wOC0xNlQxNDoyNTo0Ni4yNzE0OTA2MjNaIiwic291cmNlIjoiY29va2llIn19LCJiZGF5IjoiMjAyMy0wOC0wMlQxNDoyNTo0NS4wMzQ4MjI0OTJaIn0=',
-        },
-        url: 'https://elb.the-ozone-project.com/cdn-cgi/rum?',
-      },
-      'visitor-id': {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [12705, 12662],
-        headerType: 'request',
-        isFirstParty: false,
-        parsedCookie: {
-          domain: '.media.net',
-          expires: '2024-07-19T03:51:41.270Z',
-          httponly: false,
-          name: 'visitor-id',
-          path: '/',
-          samesite: 'none',
-          secure: true,
-          value: '3328267010240970000V10',
-        },
-        url: 'https://hbx.media.net/cksync.php?cs=1&type=pbs&ovsid=setstatuscode&bidder=medianet&gdpr=0&gdpr_consent=&us_privacy=pbs-ozone&redirect=https%3A%2F%2Felb.the-ozone-project.com%2Fsetuid%3Fbidder%3Dmedianet%26gdpr%3D0%26gdpr_consent%3D%26uid%3D%24UID',
-      },
-      xbc: {
-        analytics: {
-          category: '',
-          dataController: '',
-          description: '',
-          domain: '',
-          gdprUrl: '',
-          name: '',
-          platform: '',
-          retention: '',
-          wildcard: '',
-        },
-        frameIdList: [0],
-        headerType: 'request',
-        isFirstParty: true,
-        parsedCookie: {
-          domain: '.bbc.com',
-          expires: '2024-09-17T01:54:50.396Z',
-          httponly: false,
-          name: 'xbc',
-          path: '/',
-          samesite: 'lax',
-          secure: false,
-          value: '%7Bkpex%7DQEXSp2lkur8K7DvifXbzTt_xnr4Kckd7ir2Ie_CEavY',
-        },
-        url: 'https://gn-web-assets.api.bbc.com/assets/imgs/BBC_Storyworks_NoCP.svg',
-      },
-    },
-    focusedAt: 1691983869724,
-  },
-};
